implement basic widget module handling
implements a simple period-info widget for testing purposes.
Further widgets can be added in a similar way. If added to the
list of widgets in the gulpfile gulp will include them in
chronontology-widgets.js. This file serves as a general library
that can be included on any website to use the chronontology
widgets.

Closes #7381